requirements.txt
streamlit
pandas
openpyxl
plotly.express
qrcode